Type ediblemushroom lunge-scan angle accumulators as s16
fn_801D129C kept anglePlus/angleMinus as int, forcing an extsh at each
of the three s16 returns. The retail build holds the accumulators
sign-extended (extsh once at the getAngle result) and returns them with
a plain mr. Declaring them s16 drops the per-return extsh. 98.3 -> 98.98%.

+++ src/main/dll/dll_01A7_ediblemushroom.c
@@ -455,8 +455,8 @@ void edibleMushroomFn_801d083c(u8* obj, u8* state, u8* other)
 s16 fn_801D129C(u8* obj, u8* player, u8* state, f32 dist)
 {
     s16 angle;
-    int anglePlus;
-    int angleMinus;
+    s16 anglePlus;
+    s16 angleMinus;
     int i;
     f32 rad;
     f32 c;
